requirements.txt
boto3
python-decouple
glob2